Turn off HTTPS & IP pinning
Since we changed the names of the deployment,
we end up with new namespaces & service IPs. DNS isn't
set up properly for these yet, so we would have to turn HTTPS
off for now.

Note that this doesn't mess with any user storage, since that
stays under the same prefix as before.