[WIP] Add dask-gateway configuration
This adds a dask-gateway deployment to the dev hub. This required a
small change to the JupyterHub configuration, as well as the addition of
the dask-gateway helm chart. Things are still rough. TODO:

- [ ] Integrate as a subchart, moving gateway config into normal
`common.yaml` file.
- [ ] Update images to include new dask-gateway release
- [ ] Add default user-side `~/.config/dask/gateway.yaml` file to image
to provide a nicer user configuration (e.g. no need to enter addresses
manually).

[ci skip]